Update DOIs
Removed eLife DOIs from molecular structures and so updated rest to order. NOTE to vendor: do not reorder DOIs or change any DOIs once past first proof stage and DO NOT change post publication as the DOIs are registered. They should be arbitrary numbers and so the .000 suffix order is not import, but ordered for ease of process.
Made some other minor changes to match other version of kitchen sink exactly - formatting only